+++ _inc/blocks/sharing-button/view.asset.php
@@ -1,1 +1,1 @@
-<?php return array('dependencies' => array(), 'version' => 'd9db3cc293229ed219b5');
+<?php return array('dependencies' => array(), 'version' => '5fe0abcaac0da5b3fafa');